# services/reporting-service/openapi.yaml
openapi: 3.1.0
info:
  title: SiberIndo CTI Reporting Service API
  version: "1.0.0"
  description: >
    Report request, status tracking, and retrieval for the SiberIndo CTI platform.
    A report is requested via POST /v1/reports (created in the 'queued' state), then
    generated asynchronously: the service publishes report.requested to Kafka, its
    own worker consumes that event and transitions the report through
    'generating' -> 'complete' (or 'failed'), publishing report.completed when done.

    All responses use the platform envelope: successful responses wrap the payload in
    `{ "data": ..., "meta": { "request_id", "timestamp" } }`; errors return
    `{ "error": { "code", "message", "details" }, "meta": { "request_id", "timestamp" } }`.
servers:
  - url: /
security:
  - bearerAuth: []
tags:
  - name: reports
paths:
  /v1/reports:
    get:
      tags: [reports]
      summary: List reports
      description: Returns the tenant's reports, newest first. Requires the `report:read` permission.
      parameters:
        - name: report_type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ReportType'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ReportStatus'
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        '200':
          description: A list of reports.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Report' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [reports]
      summary: Request a report
      description: >
        Queues a new report and publishes report.requested for asynchronous
        generation. Requires the `report:create` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateReportRequest' }
      responses:
        '201':
          description: The queued report.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Report' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/reports/{id}:
    get:
      tags: [reports]
      summary: Get a report
      description: Returns a single report by id, including its current generation status. Requires the `report:read` permission.
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: The report.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Report' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/reports/{id}/download:
    get:
      tags: [reports]
      summary: Get a completed report's output reference
      description: >
        Returns the storage reference for a completed report. If the report has not
        finished generating, responds 409 CONFLICT. Requires the `report:read`
        permission.
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: The output reference and status of a completed report.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          output_ref: { type: string, example: "s3://siberindo-reports/<tenant>/<report>.pdf" }
                          status: { type: string, example: complete }
                        required: [output_ref, status]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The report is not yet complete.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ReportId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    ReportType:
      type: string
      enum:
        - executive_summary
        - module_finding_report
        - takedown_status_report
        - credential_exposure_report
        - threat_actor_report
        - sla_compliance_report
    ReportStatus:
      type: string
      enum: [queued, generating, complete, failed]
    OutputFormat:
      type: string
      enum: [pdf, csv, json, xlsx]
      default: pdf
    CreateReportRequest:
      type: object
      required: [report_type, title]
      properties:
        report_type: { $ref: '#/components/schemas/ReportType' }
        title: { type: string, maxLength: 512 }
        output_format: { $ref: '#/components/schemas/OutputFormat' }
        parameters:
          type: object
          description: Free-form report parameters stored as JSONB. Defaults to {}.
          additionalProperties: true
    Report:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        report_type: { $ref: '#/components/schemas/ReportType' }
        title: { type: string }
        status: { $ref: '#/components/schemas/ReportStatus' }
        parameters:
          type: object
          additionalProperties: true
        output_ref:
          type: string
          nullable: true
        output_format: { $ref: '#/components/schemas/OutputFormat' }
        requested_by:
          type: string
          format: uuid
          nullable: true
        generated_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        failure_reason:
          type: string
          nullable: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, tenant_id, report_type, title, status, output_format, created_at, updated_at]
    Meta:
      type: object
      properties:
        request_id: { type: string }
        timestamp: { type: string, format: date-time }
      required: [request_id, timestamp]
    SuccessEnvelope:
      type: object
      properties:
        data: {}
        meta: { $ref: '#/components/schemas/Meta' }
      required: [data, meta]
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - RESOURCE_NOT_FOUND
                - CONFLICT
                - BUSINESS_RULE_VIOLATION
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
            message: { type: string }
            details: {}
          required: [code, message]
        meta: { $ref: '#/components/schemas/Meta' }
      required: [error, meta]
  responses:
    ValidationError:
      description: The request failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: The token lacks the required permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: The report does not exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }